requirements.txt
numpy==1.18.5
Pillow==7.1.2